实现了controller层；更新了service层中内容

+++ src/main/java/com/fudanscetool/springboot/pojo/Stage2.java
@@ -18,5 +18,5 @@ public class Stage2 {
     private String FCIL;
     private String SCED;
 
-    private float estimateLoad;
+    private double estimateLoad;
 }
